unpack widget
