+++ modules/executors/other/unpack_parser.py
@@ -1,0 +1,5 @@
+from modules.executors.common import BaseGuiParser
+from .copy_parser import CopyParser
+
+class UnpackParser(CopyParser):
+    node_types = ["copy_unpack"]
